Vuelvo a double en algunas cosas por las dudas. Agua1 tiene una geometria fea ahora.

+++ g2g/cpu/pot.cpp
@@ -39,7 +39,7 @@ using namespace std;
 
 using namespace G2G;
 
-void cpu_pot(float dens, float& ex, float& ec, float& y2a)
+void cpu_pot(float dens, double& ex, double& ec, double& y2a)
 {
 	// data X alpha
 
@@ -140,7 +140,7 @@ static void gcorc(double rtrs, double& gg, double& grrs);
  9: PBE
 */
 
-void cpu_potg(float dens, const float3& grad, const float3& hess1, const float3& hess2, float& ex, float& ec, float& y2a)
+void cpu_potg(float dens, const double3& grad, const double3& hess1, const double3& hess2, double& ex, double& ec, double& y2a)
 {
   // hess1: xx, yy, zz
   // hess2: xy, xz, yz
